Foundation
Lots of work has been done, almost all functional components are stateful and operate as designed. Lots of inactive buttons at this time, but I believe if I hosted the site on Firebase and used axios to fetch data I could have this site up and running in no time!